Added extended range version of splash potion and many new configuration options
Splash Potion of Teleportation II has a splash range of 7x3x7 compared to the normal version range of 4x2x4. Range is now included in the potion tooltip.
Wands and beacons can now be crafted with an Eye of Ender in addition to an Ender Pearl.
Added crafting difficulty config setting where HARD requires an Eye of Ender in the wand and beacon recipes. To facilitate this, a RecipeTeleportationItem class was created for an IRecipeFactory, which is now referenced in the json recipe files for the wand and beacon, through a _factories.json resource file.
Added 9 splash potion config options to fine tune what in the splash effect radius will be teleported.
Added beacon tooltip.