Teleport beacons can become teleport pads with stored destinations
Gave teleport beacons the same ability as teleport rails to store a teleport destination. Left-click them with a wand of teleportation to toggle their direction between sender (aka teleport pad) and receiver (their default). Shift+right-click with a wand to store a receiver beacon from your network as a destination.
Teleport pads will teleport ANY entity that's on top of it if it has a stored destination.
Removed the IS_ACTIVE blockstate from beacons. Changed it to a tile entity field called isStored, same as teleport rails.