Bug fixes, sound changes, and other refactoring
Change teleport block variable isStored to be client-only (it never should have been set there) and stop writing it to NBT. Also, created getter and setter for it in ITeleportationBlock interface for convenience by the wand.
Fixed an infinite loop bug in the wand shift+right-click of a teleport pad, found during playtesting.
Teleport rail placement and storage-toggling sounds are now the same as beacon.