Improving player messages, some refactoring, and other minor changes
Added TeleportDirection as a stored field in TeleportDestination. This provides more accurate information for the player.
Refactored teleport block name getters and setters (e.g. getBeaconName() and getRailName()) into the interface method getTeleportName() and setTeleportName().
Refactored the code for displaying a beacon or rail name into the TeleportationHelper method displayTeleportBlockName().
Standardized and improved many of the names of the language entries and the messages to the player about teleport blocks, to improve consistency and provide more detail.
Added more tip lines for the teleport wand and blocks for completeness.