Added Bow of Teleportation and storage of destination in teleportation projectiles
New item: Bow of Teleportation.
All splash potions and arrows of teleportation now read and store the appropriate teleport destination when their entity is constructed.
TeleportationHandler now has a specialDestination field which could be multipurpose but is currently only used by the player's teleportation handler for storing the destination for the Bow of Teleportation.
Fixed another bug in TeleportationHandler.getNextDestination (index out of range).